+++ onsiteprint-plugin/blocks/event/block-script-parts/participant/get-participant.js
@@ -1,0 +1,70 @@
+/* ------------------------------------------------------------------------
+ #  JS Part Name: Get Participant
+ *  Getting the Participant from the Participant List in the Event Block.
+ ?  Updated: 2025-05-22 - 03:42 (Y:m:d - H:i)
+ ?  Info: Added new Function, opGetParticipant().
+---------------------------------------------------------------------------
+ #  TABLE OF CONTENTS:
+---------------------------------------------------------------------------
+
+	1. 	Import Functions from Scripts
+
+    2. 	Function: Get the Participant from the Participant List
+
+---------------------------------------------------------------------------
+ #  1. Import Functions from Scripts
+--------------------------------------------------------------------------- */
+import * as opModuleBasic from '../../../../assets/js/inc/basic.js'
+import * as opModuleEvent from '../../../../assets/js/inc/event/event.js'
+
+/* ------------------------------------------------------------------------
+ #  2. Function: Get the Participant from the Participant List
+--------------------------------------------------------------------------- */
+export function opGetParticipant( debug, participantId ) {
+
+    try {
+
+        ///// Get Function Name.
+        var functionName = opGetParticipant.name
+
+        ///// Set the Debug.
+        ////* Set the Parameter If is not defined (true or false).
+        if ( debug !== true ) debug = false
+        if ( debug ) console.group( `${ functionName }()` )
+
+
+
+
+
+
+
+        ///// Return the Response.
+        return opModuleBasic.opReturnResponse( false, 200, { 
+            message: `The Participant was Found!`, 
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        }, debug )
+
+    } catch( errorResponse ) {
+
+        ///// Create Error Details.
+        let errorDetails = ( errorResponse.error == true ) ? errorResponse : opModuleBasic.opReturnResponse( true, 400, { 
+            message: errorResponse.message,
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        } )
+
+        ///// Log Error Details in the Console.
+        if ( debug ) console.error( 'ERROR:', errorDetails )
+
+        ///// Return the Error Response.
+        return errorDetails
+
+    } finally {
+
+        ///// End the Console Log Group.
+        if ( debug ) console.groupEnd()
+
+    }
+
+}