+++ onsiteprint-plugin/blocks/event/block-script-parts/parts.js
@@ -1,8 +1,8 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Parts
  *  Block functions included in the Event Block Script.
- ?  Updated: 2025-04-14 - 03:51 (Y:m:d - H:i)
- ?  Info: Added new Function (opAddParticipant).
+ ?  Updated: 2025-05-22 - 03:42 (Y:m:d - H:i)
+ ?  Info: Added new Function, opGetParticipant().
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
 ---------------------------------------------------------------------------
@@ -10,6 +10,8 @@
     1. 	Setup List
 
     2.  Add Participant
+
+    2.  Get Participant
 
 ---------------------------------------------------------------------------
  #  1. Setup List
@@ -22,3 +24,9 @@ export { opSetupList }
 --------------------------------------------------------------------------- */
 import { opAddParticipant } from './participant/add-participant.js'
 export { opAddParticipant }
+
+/* ------------------------------------------------------------------------
+ #  3. Get Participant
+--------------------------------------------------------------------------- */
+import { opGetParticipant } from './participant/get-participant.js'
+export { opGetParticipant }